Changed progress bar in fragments. Added Rens fragments
