+++ app/src/main/java/grp02/brg_app/View/BroegActivity1.java
@@ -12,7 +12,7 @@ import android.view.MenuItem;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
 import grp02.brg_app.R;
-import grp02.brg_app.View.Fragments.NameStart;
+import grp02.brg_app.View.BrygFragments.NameStart;
 
 
 public class BroegActivity1 extends AppCompatActivity{
