+++ app/src/main/java/grp02/brg_app/View/BrygFragments/OnPressedBryg.java
@@ -1,6 +1,5 @@
-package grp02.brg_app.View.Fragments;
+package grp02.brg_app.View.BrygFragments;
 
-import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
@@ -8,8 +7,6 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
-import java.util.concurrent.TimeUnit;
 
 import grp02.brg_app.R;
