+++ app/src/main/res/layout/activity_rens1.xml
@@ -7,6 +7,16 @@
     android:background="@color/colorAccent"
     tools:context=".RensActivity1">
 
+    <FrameLayout
+        android:id="@+id/rensFragment"
+        android:layout_width="409dp"
+        android:layout_height="380dp"
+        app:layout_constraintBottom_toTopOf="@+id/bottomNav"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/headerImage"
+        app:layout_constraintVertical_bias="1.0" />
+
     <View
         android:layout_width="match_parent"
         android:layout_height="1dp"
