+++ app/src/main/res/layout/fragment_grind_size.xml
@@ -6,17 +6,46 @@
     android:layout_height="match_parent"
     android:background="#E0E0E0">
 
+    <Button
+        android:id="@+id/buttonNextGrind"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="285dp"
+        android:layout_marginEnd="73dp"
+        android:background="@drawable/default_btn"
+        android:drawableEnd="@drawable/ic_baseline_navigate_next_24"
+        android:text="@string/n_ste"
+        android:textColor="@color/colorPrimary"
+        app:layout_constraintBottom_toTopOf="@+id/progressBar"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="1.0" />
+
     <ProgressBar
         android:id="@+id/progressBarGrind"
         style="?android:attr/progressBarStyleHorizontal"
         android:layout_width="266dp"
         android:layout_height="25dp"
+        android:layout_marginTop="333dp"
         android:progressDrawable="@drawable/default_processbar"
-        android:layout_marginTop="104dp"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.496"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/scroll_choice_grind" />
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <Button
+        android:id="@+id/buttonTilbageGrind"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="73dp"
+        android:layout_marginTop="285dp"
+        android:background="@drawable/default_btn"
+        android:drawableStart="@drawable/ic_baseline_arrow_back_ios_24"
+        android:text="@string/tilbage"
+        android:textColor="@color/colorPrimary"
+        app:layout_constraintBottom_toTopOf="@+id/progressBar"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="1.0" />
 
     <TextView
         android:id="@+id/textViewGrind"
@@ -51,27 +80,6 @@
         app:scroll_item_text_size="16dp"
         app:scroll_selected_item_text_color="#8A6F62"
         app:scroll_visible_item_count="7" />
-
-    <Button
-        android:id="@+id/buttonNextGrind"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/n_ste"
-        android:drawableEnd="@drawable/ic_baseline_navigate_next_24"
-        android:background="@drawable/default_btn"
-        android:textColor="@color/colorPrimary"
-        app:layout_constraintBottom_toTopOf="@+id/progressBarGrind"
-        app:layout_constraintEnd_toEndOf="@+id/progressBarGrind" />
-    <Button
-        android:id="@+id/buttonTilbageGrind"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/tilbage"
-        android:drawableStart="@drawable/ic_baseline_arrow_back_ios_24"
-        android:background="@drawable/default_btn"
-        android:textColor="@color/colorPrimary"
-        app:layout_constraintBottom_toTopOf="@+id/progressBarGrind"
-        app:layout_constraintStart_toStartOf="@+id/progressBarGrind" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
