+++ app/src/main/res/layout/fragment_on_save_bryg.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.Fragments.OnSaveBryg">
+    tools:context=".View.BrygFragments.OnSaveBryg">
 
     <com.airbnb.lottie.LottieAnimationView
         android:layout_width="300dp"
