+++ app/src/main/res/layout/fragment_start.xml
@@ -6,41 +6,11 @@
     android:layout_height="match_parent"
     android:background="#E0E0E0">
 
-    <TextView
-        android:id="@+id/textViewStart"
-        android:layout_width="276dp"
-        android:layout_height="77dp"
-        android:layout_marginTop="12dp"
-        android:background="#E0E0E0"
-        android:gravity="center"
-        android:text="Hvad vil du gerne kalde din brøg?"
-        android:textColor="#77482C"
-        android:textColorLink="#77482C"
-        android:textSize="20dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.495"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/buttonNextStart"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="215dp"
-        android:background="@drawable/default_btn"
-        android:drawableEnd="@drawable/ic_baseline_navigate_next_24"
-        android:text="@string/n_ste"
-        android:textColor="@color/colorPrimary"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.761"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textViewStart" />
-
     <EditText
         android:id="@+id/navnPaaBroeggen"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="20dp"
+        android:layout_marginTop="16dp"
         android:ems="10"
         android:inputType="textPersonName"
         android:text=""
@@ -50,5 +20,36 @@
         app:layout_constraintHorizontal_bias="0.497"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textViewStart" />
+
+    <TextView
+        android:id="@+id/textViewStart"
+        android:layout_width="276dp"
+        android:layout_height="77dp"
+        android:layout_marginTop="8dp"
+        android:background="#E0E0E0"
+        android:gravity="center"
+        android:text="Hvad vil du gerne kalde din brøg?"
+        android:textColor="#77482C"
+        android:textColorLink="#77482C"
+        android:textSize="20dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.496"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <Button
+        android:id="@+id/buttonNextStart"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="285dp"
+        android:layout_marginEnd="74dp"
+        android:background="@drawable/default_btn"
+        android:drawableEnd="@drawable/ic_baseline_navigate_next_24"
+        android:text="@string/n_ste"
+        android:textColor="@color/colorPrimary"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="1.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
